W172l: repaired the struct-migration blast radius + fclCombineMisc FOUR window clears.
REPAIR: brp_birth.c was COMPILE_ERROR on all 10 of its markers. The new SflResourceManager header correctly retyped sflResInit's parameter, but brp_birth still passed a raw u8*. Added the same cast br_root already uses; brp_birth is back to 9/10 MATCH with ZERO metric changes, and a repo-wide grep confirms only two sflResInit call sites exist and both are now migrated. Caught by the hygiene-audit agent, which correctly reported rather than repaired it. Lesson for the struct layer: retyping a shared prototype REQUIRES sweeping every caller in the same pass, and a full-tree compile check is the only way to prove it.

fclCombineMisc 003D6200 744/672 nd458 -> 660/672 nd333, 003D7AC0 and 003D7DA0 both 748/736 nd521 -> 704/736 nd399 (true twins), 003D8630 536/512 nd379 -> 500/512 nd338 -- all four by copying the matching 003D5BB0 base-first rule-table idiom (one DAT_006a5410 record base with +4/+6 fields), removing redundant boolean carry state in the twin selectors, converting their exit-at-loop-head shape to a bounded for loop, and recovering a simple signed-short comparison.
fclShop: five raw assertion-string literals replaced with the existing typed DAT_006a6af8 declaration (003E6E80 nd3367 -> 3362, 003E8860 nd988 -> 974).
NEGATIVE recorded: fclShop 003ED5E0 -- retail's early-return guard layout plus separating the reused state/detail pointer role regressed nd714 -> 1088 at unchanged size; fully reverted to baseline.

Slice over-window 13 -> 5, zero lost, zero additions, 241 markers unchanged, pragma brackets unchanged.

+++ src/h_pad.c
@@ -422,7 +422,7 @@ void H_Pad_UpdateRumble(void)
             goto rumble_off;
         }
         {
-            s32 intensity;
+            u16 intensity;
             u8 intensityByte;
             intensityByte = sRumbleIntensity.b;
             intensity = intensityByte;
